Get top podcasts from PodcastManager instead of EpisodeManager
